fix: Use correct digest for SCRAM channel binding
Currently, when connecting to a server using a certificate with a
SHA2-384-based signature algorithm in "sslmode=require", the connection
fails with the following error:
"FATAL: SCRAM channel binding check failed".

This is because the client is hard-coded to use the SHA2-256 digest
algorithm to calculate the channel binding. The digest algorithm used
should be derived from the server's certificate.

This change parses the server's certificate, extracting the digest
algorithm and using that to calculate the channel binding.